move error dialog folders to /pages/common
