+++ src/app/pages/common/error-dialog/error.dialog.scss
@@ -1,0 +1,22 @@
+@import '~assets/scss/msw-alert-dialog.scss';
+
+.container {
+  width: unset;
+}
+
+.bottomline {
+  height: unset;
+}
+
+.details {
+  width: fit-content !important;
+  min-width: 380px;
+  background: rgba(0, 0, 0, 0.01);
+  border: 0.5px solid rgba(0, 0, 0, 0.03);
+  border-radius: 5px;
+  display: block;
+  margin: 16px auto;
+  overflow-x: auto;
+  padding: 20px;
+  white-space: pre-wrap;
+}
